+++ src/layers/XRLayer/shader-modules/index.js
@@ -1,0 +1,4 @@
+import channels from './channel-intensity-module';
+import lens from './lens-module';
+
+export { lens, channels };