+++ src/layers/XRLayer/shader-modules/lens-module.js
@@ -1,0 +1,9 @@
+import fs from './lens.glsl';
+
+export default {
+  name: 'lens-module',
+  defines: {
+    SAMPLER_TYPE: 'usampler2D'
+  },
+  fs
+};