+++ src/layers/XRLayer/xr-layer-fragment-colormap.webgl2.glsl
@@ -2,68 +2,21 @@
 #define SHADER_NAME xr-layer-fragment-shader
 precision highp float;
 precision highp int;
-precision highp usampler2D;
-
-#pragma glslify: jet = require("glsl-colormap/jet")
-#pragma glslify: hsv = require("glsl-colormap/hsv")
-#pragma glslify: hot = require("glsl-colormap/hot")
-#pragma glslify: cool = require("glsl-colormap/cool")
-#pragma glslify: spring = require("glsl-colormap/spring")
-#pragma glslify: summer = require("glsl-colormap/summer")
-#pragma glslify: autumn = require("glsl-colormap/autumn")
-#pragma glslify: winter = require("glsl-colormap/winter")
-#pragma glslify: bone = require("glsl-colormap/bone")
-#pragma glslify: copper = require("glsl-colormap/copper")
-#pragma glslify: greys = require("glsl-colormap/greys")
-#pragma glslify: yignbu = require("glsl-colormap/yignbu")
-#pragma glslify: greens = require("glsl-colormap/greens")
-#pragma glslify: yiorrd = require("glsl-colormap/yiorrd")
-#pragma glslify: bluered = require("glsl-colormap/bluered")
-#pragma glslify: rdbu = require("glsl-colormap/rdbu")
-#pragma glslify: picnic = require("glsl-colormap/picnic")
-#pragma glslify: rainbow = require("glsl-colormap/rainbow")
-#pragma glslify: portland = require("glsl-colormap/portland")
-#pragma glslify: blackbody = require("glsl-colormap/blackbody")
-#pragma glslify: earth = require("glsl-colormap/earth")
-#pragma glslify: electric = require("glsl-colormap/electric")
-#pragma glslify: alpha = require("glsl-colormap/alpha")
-#pragma glslify: viridis = require("glsl-colormap/viridis")
-#pragma glslify: inferno = require("glsl-colormap/inferno")
-#pragma glslify: magma = require("glsl-colormap/magma")
-#pragma glslify: plasma = require("glsl-colormap/plasma")
-#pragma glslify: warm = require("glsl-colormap/warm")
-#pragma glslify: rainbow-soft = require("glsl-colormap/rainbow-soft")
-#pragma glslify: bathymetry = require("glsl-colormap/bathymetry")
-#pragma glslify: cdom = require("glsl-colormap/cdom")
-#pragma glslify: chlorophyll = require("glsl-colormap/chlorophyll")
-#pragma glslify: density = require("glsl-colormap/density")
-#pragma glslify: freesurface-blue = require("glsl-colormap/freesurface-blue")
-#pragma glslify: freesurface-red = require("glsl-colormap/freesurface-red")
-#pragma glslify: oxygen = require("glsl-colormap/oxygen")
-#pragma glslify: par = require("glsl-colormap/par")
-#pragma glslify: phase = require("glsl-colormap/phase")
-#pragma glslify: salinity = require("glsl-colormap/salinity")
-#pragma glslify: temperature = require("glsl-colormap/temperature")
-#pragma glslify: turbidity = require("glsl-colormap/turbidity")
-#pragma glslify: velocity-blue = require("glsl-colormap/velocity-blue")
-#pragma glslify: velocity-green = require("glsl-colormap/velocity-green")
-#pragma glslify: cubehelix = require("glsl-colormap/cubehelix")
-
+precision highp SAMPLER_TYPE;
 
 // our texture
-uniform usampler2D channel0;
-uniform usampler2D channel1;
-uniform usampler2D channel2;
-uniform usampler2D channel3;
-uniform usampler2D channel4;
-uniform usampler2D channel5;
+uniform SAMPLER_TYPE channel0;
+uniform SAMPLER_TYPE channel1;
+uniform SAMPLER_TYPE channel2;
+uniform SAMPLER_TYPE channel3;
+uniform SAMPLER_TYPE channel4;
+uniform SAMPLER_TYPE channel5;
 
 // range
 uniform vec2 sliderValues[6];
 
 // opacity
 uniform float opacity;
-uniform float divisor;
 
 in vec2 vTexCoord;
 
@@ -72,19 +25,19 @@ out vec4 color;
 
 
 void main() {
-  float intensityValue0 = (float(texture(channel0, vTexCoord).r) - sliderValues[0][0]) / max(0.0005, (sliderValues[0][1] - sliderValues[0][0]));
-  float intensityValue1 = (float(texture(channel1, vTexCoord).r) - sliderValues[1][0]) / max(0.0005, (sliderValues[1][1] - sliderValues[1][0]));
-  float intensityValue2 = (float(texture(channel2, vTexCoord).r) - sliderValues[2][0]) / max(0.0005, (sliderValues[2][1] - sliderValues[2][0]));
-  float intensityValue3 = (float(texture(channel3, vTexCoord).r) - sliderValues[3][0]) / max(0.0005, (sliderValues[3][1] - sliderValues[3][0]));
-  float intensityValue4 = (float(texture(channel4, vTexCoord).r) - sliderValues[4][0]) / max(0.0005, (sliderValues[4][1] - sliderValues[4][0]));
-  float intensityValue5 = (float(texture(channel5, vTexCoord).r) - sliderValues[5][0]) / max(0.0005, (sliderValues[5][1] - sliderValues[5][0]));
+  float intensityValue0 = sample_and_apply_sliders(channel0, vTexCoord, sliderValues[0]);
+  float intensityValue1 = sample_and_apply_sliders(channel1, vTexCoord, sliderValues[1]);
+  float intensityValue2 = sample_and_apply_sliders(channel2, vTexCoord, sliderValues[2]);
+  float intensityValue3 = sample_and_apply_sliders(channel3, vTexCoord, sliderValues[3]);
+  float intensityValue4 = sample_and_apply_sliders(channel4, vTexCoord, sliderValues[4]);
+  float intensityValue5 = sample_and_apply_sliders(channel5, vTexCoord, sliderValues[5]);
 
   float intensityArray[6] = float[6](intensityValue0, intensityValue1, intensityValue2, intensityValue3, intensityValue4, intensityValue5);
   float intensityCombo = 0.0;
   for(int i = 0; i < 6; i++) {
     intensityCombo += max(0.0,intensityArray[i]);
   }
-  color = vec4(colormapFunction(min(1.0,intensityCombo)).xyz, opacity);
+  color = colormap(intensityCombo, opacity);
   geometry.uv = vTexCoord;
   DECKGL_FILTER_COLOR(color, geometry);
 }