+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - Fix snackbars for errors and initial image.
 - Export custom zarr `HTTPStore` with abort controller signal support.
+- Modularize shaders using deck.gl shader modules.
 - Provide pure ESM export of Viv and use separate build for Avivator
 - Downgrade react back to 16.8